build: Fix Husky setup
The Husky setup wasn't working correctly, at least for WSL setups. This
changes the Husky set up to follow the newest way of doing it, which
seems to work in WSL.

Updates Husky to newest version. Apparently hooks are no longer set in
package.json, but instead in a bash script in the .husky folder.

Updates the pre-commit hook to also run npm test non-interactively