feat: Add event details page
Adds a link from each event card to its details page.

Adds an event details page that lists more information about the event
as well as a list of matches in the event.

Adds a MatchCard component to display info about a match. Adds a
MatchCards component to render a set of MatchCards. Adds test for both
of these components.

Adds a route for the event details page