fix: DataGridExample
To work with newer `@material-ui/data-grid` versions:

Allow valueGetter function params to be inferred so that we ensure it
gets appopriately typed to the updated type from GridColDef
automatically. The new type is GridValueGetterParams.

Use hook provided by the package to get the props for the Toolbar grid
slot component, since the passed in props were removed in version
v4.0.0-alpha.24.